Updated jssc to be compatible with maven,osgi and to compile with maven. Still missing is a custom loader for native libraries. And a pom.xml which will build the natives.
