+++ src/main/cpp/_nix_based/jssc.cpp
