+++ src/main/cpp/jssc_SerialNativeInterface.h
