+++ src/main/cpp/windows/jssc.c++
