+++ src/main/java/org/scream3r/jssc/SerialPortEvent.java
@@ -22,7 +22,7 @@
  * e-mail: scream3r.org@gmail.com
  * web-site: http://scream3r.org | http://code.google.com/p/java-simple-serial-connector/
  */
-package jssc;
+package org.scream3r.jssc;
 
 /**
  *
