+++ src/main/java/org/scream3r/jssc/SerialPortEventListener.java
@@ -22,43 +22,13 @@
  * e-mail: scream3r.org@gmail.com
  * web-site: http://scream3r.org | http://code.google.com/p/java-simple-serial-connector/
  */
-package jssc;
+package org.scream3r.jssc;
 
 /**
  *
  * @author scream3r
  */
-public class SerialPortTimeoutException extends Exception {
+public interface SerialPortEventListener {
 
-    private String portName;
-    private String methodName;
-    private int timeoutValue;
-
-    public SerialPortTimeoutException(String portName, String methodName, int timeoutValue) {
-        super("Port name - " + portName + "; Method name - " + methodName + "; Serial port operation timeout (" + timeoutValue + " ms).");
-        this.portName = portName;
-        this.methodName = methodName;
-        this.timeoutValue = timeoutValue;
-    }
-
-    /**
-     * Getting port name during operation with which the exception was called
-     */
-    public String getPortName(){
-        return portName;
-    }
-
-    /**
-     * Getting method name during execution of which the exception was called
-     */
-    public String getMethodName(){
-        return methodName;
-    }
-
-    /**
-     * Getting timeout value in millisecond
-     */
-    public int getTimeoutValue(){
-        return timeoutValue;
-    }
+    public abstract void serialEvent(SerialPortEvent serialPortEvent);
 }
