+++ src/main/resources/libs/linux/libjSSC-2.8_armhf.so
