+++ src/main/resources/libs/linux/libjSSC-2.8_x86_64.so
