+++ src/main/resources/libs/mac_os_x/libjSSC-2.8_x86.jnilib
