+++ src/main/resources/libs/solaris/libjSSC-2.8_x86.so
