+++ src/main/resources/libs/solaris/libjSSC-2.8_x86_64.so
